+++ common/src/main/java/com/zigythebird/playeranim/animation/keyframe/Keyframe.java
@@ -22,12 +22,12 @@ import java.util.Objects;
  * @param easingType The {@code EasingType} to use for transformations
  * @param easingArgs The arguments to provide to the easing calculation
  */
-public record Keyframe<T extends MolangExpression>(double length, T startValue, T endValue, EasingType easingType, List<T> easingArgs) {
-	public Keyframe(double length, T startValue, T endValue) {
+public record Keyframe<T extends MolangExpression>(float length, T startValue, T endValue, EasingType easingType, List<T> easingArgs) {
+	public Keyframe(float length, T startValue, T endValue) {
 		this(length, startValue, endValue, EasingType.LINEAR);
 	}
 
-	public Keyframe(double length, T startValue, T endValue, EasingType easingType) {
+	public Keyframe(float length, T startValue, T endValue, EasingType easingType) {
 		this(length, startValue, endValue, easingType, new ObjectArrayList<>(0));
 	}
